\name{draw_i}
\alias{draw_i}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
generate random integer numbers
}
\description{
Random +1/-1 numbers. This is a secondary function used by bin_ps
}
\usage{
draw_i()
}
%- maybe also 'usage' for other objects documented here.
\value{
+1,-1
}
\author{
Francisco Juretig <fjuretig@yahoo.com>
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
